# Multiple performance improvements