@berenddeboer 2021-09-16
This fix is remarkably effective. It cuts down my processing time in half.

I'm reading 102 smallish files (about 1MB each) from S3, and writing to AWS Aurora Serverless. I'm writing about 200,00 records. This patch cuts down my time from 6 minutes to 3 minutes.