@berenddeboer 2021-09-16
And another test: 8.8 million rows. Used to take 86 minutes, now down to 57 (while the machine was overloaded doing other stuff). I think this should really be merged!