@laurentS 2022-04-04
FWIW, I've forked this repo and added some further performance improvements and memory savings in https://github.com/oviohub/target-postgres/